Extract lifecycle_hooks.go and lifecycle_interfaces.go from interfaces.go
